fix the visit entity errors

+++ src/main/java/com/projects/shortify_backend/dto/response/RedirectResponseDTO.java
@@ -13,7 +13,7 @@ public class RedirectResponseDTO {
 
     private String device;
 
-    private Long numberOfVisit;
+    private Integer numberOfVisit;
 
     private String originalUrl;
