+++ src/main/java/com/projects/shortify_backend/dto/response/base/VisitorBaseResponseDTO.java
@@ -15,5 +15,5 @@ public class VisitorBaseResponseDTO {
 
     private String device;
 
-    private Long numberOfClicked;
+    private Integer numberOfClicked;
 }
